requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
opencv-python==4.5.3.56
dlib==19.22.0
numpy==1.21.2
Pillow==8.3.2
flask-login==0.5.0
flask-wtf==0.15.1
werkzeug==2.0.1
pytz==2021.1
requests==2.26.0
flask-migrate==3.1.0
flask-cors==3.0.10
python-dotenv==0.19.0